+++ src/dye/utils.py
@@ -24,6 +24,8 @@
 from importlib import metadata
 
 import benedict
+import rich
+
 
 from .exceptions import DyeSyntaxError
 
@@ -53,9 +55,14 @@ def benedict_keylist(d):
     return [".".join([f"{key}" for key in kl]) for kl in kls]
 
 
-def colors_merge_and_process(base_colors, merging_colors, jinja_env):
+def merge_and_process_colors(base_colors, merging_colors, jinja_env):
     """merge together two benedicts of colors (one or both can be empty) and
     process references and templates in all the values.
+
+    Args:
+        base_colors (benedict): _description_
+        merging_colors (benedict): _description_
+        jinja_env (jinja2.environments): _description_
 
     base_colors will be modified to include all of the items in merging_colors.
 
@@ -67,7 +74,7 @@ def colors_merge_and_process(base_colors, merging_colors, jinja_env):
     you just need to process one dict, call this like
 
         processed_dict = benedict()
-        colors_merge_and_process(processed_dict, unprocessed_dict)
+        merge_and_process_colors(processed_dict, unprocessed_dict)
 
     that will result in processed_dict having all the keys of unprocessed_dict
     but all the values will have been processed
@@ -96,3 +103,59 @@ def colors_merge_and_process(base_colors, merging_colors, jinja_env):
             pass
         else:
             raise DyeSyntaxError(f"color {key} must be defined as a string")
+
+
+def merge_and_process_styles(base_styles, merging_styles, jinja_env, colors=None):
+    """Merge together two benedicts of colors (one or both can be empty) and
+    process references and templates in all the values.
+
+    Args:
+        base_styles (benedict): _description_
+        merging_styles (benedict): _description_
+        jinja_env (jinja2.environments): _description_
+        colors (benedicty): optional benedict of processed colors
+
+    base_styles will be modified to include all of the items in merging_styles.
+
+    items in merging_styles will overwrite items already in base_styles.
+
+    Used by Theme and Pattern objects
+
+    This assumes that base_styles (if present) has already been processed. If
+    you just need to process one dict, call this like
+
+        processed_dict = benedict()
+        merge_and_process_styles(processed_dict, unprocessed_dict)
+
+    that will result in processed_dict having all the keys of unprocessed_dict
+    but all the values will have been processed
+    """
+    if not colors:
+        colors = benedict.benedict()
+
+    keylist = benedict_keylist(merging_styles)
+    for key in keylist:
+        value = merging_styles[key]
+        if isinstance(value, str):
+            if value in base_styles:
+                # bare lookup so that foreground_low = "foreground" works
+                base_styles[key] = base_styles[value]
+            else:
+                template = jinja_env.from_string(value)
+                rendered = template.render(
+                    # allow {{style.foreground}} or {{styles.foreground}}
+                    # or {{color.background}} or {{colors.background}}
+                    color=colors,
+                    colors=colors,
+                    style=base_styles,
+                    styles=base_styles,
+                )
+                base_styles[key] = rich.style.Style.parse(rendered)
+        elif isinstance(value, dict):
+            # So we rely on and test for the fact that both the subtable name
+            # (with the dict value) and each of the subtable keys show up in
+            # keylist. Those other keys will be picked up in a different iteration
+            # of the loop. But they shouldn't be a syntax error.
+            pass
+        else:
+            raise DyeSyntaxError(f"theme {key} must be defined as a string")